refactor : refactor component naming refs :  MBM-05-refactor-react-components
