+++ reactfront/src/App.js
@@ -4,13 +4,13 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import NavBar from "./components/UI/NavBar";
 import Footer from "./components/UI/Footer";
 //Components
-import AdminVecinos from "./components/Admin/Index/AdminVecinos";
-import CreateVecino from "./components/Admin/Crud/CreateVecino";
-import EditVecino from "./components/Admin/Crud/EditVecino";
-import ShowVecinos from "./components/Home/ShowVecinos";
-import UserMails from "./components/Mails/UserMails";
-import CreateCarta from "./components/Mails/CreateCarta";
-import LoginVecino from "./components/Login/LoginVecino";
+import AdminIndex from "./components/Admin/Index/AdminIndex";
+import CreateUser from "./components/Admin/Crud/CreateUser";
+import EditUser from "./components/Admin/Crud/EditUser";
+import Home from "./components/Home/Home";
+import ShowUserMails from "./components/Mails/ShowUserMails";
+import CreateMail from "./components/Mails/CreateMail";
+import Login from "./components/Login/Login";
 import { RequireAuth } from "react-auth-kit";
 import axios from "axios";
 import { useEffect } from "react";
@@ -20,26 +20,26 @@ function App() {
     <div className="App">
       <NavBar />
         <Routes>
-          <Route path="/" element={<ShowVecinos />}></Route>
-          <Route path="/*" element={<ShowVecinos />}></Route>
+          <Route path="/" element={<Home />}></Route>
+          <Route path="/*" element={<Home />}></Route>
           <Route
             path="/admin"
             element={
               <RequireAuth loginPath="/login">
-                <AdminVecinos />
+                <AdminIndex />
               </RequireAuth>
             }
           ></Route>
-          <Route path="/create" element={<CreateVecino />}></Route>
-          <Route path="/edit/:id" element={<EditVecino />}></Route>
+          <Route path="/create" element={<CreateUser />}></Route>
+          <Route path="/edit/:id" element={<EditUser />}></Route>
 
-          <Route path="/vecinoCartas/:id" element={<UserMails />}></Route>
+          <Route path="/vecinoCartas/:id" element={<ShowUserMails />}></Route>
           <Route path="/vecinoCartas/:id" element={<checkAuth />}></Route>
           {/* CARTAS */}
-          <Route path="/createCarta/:id" element={<CreateCarta />}></Route>
+          <Route path="/createCarta/:id" element={<CreateMail />}></Route>
           {/* Login /register */}
-          <Route path="/login" element={<LoginVecino />}></Route>
-          <Route path="/register" element={<CreateVecino />}></Route>
+          <Route path="/login" element={<Login />}></Route>
+          <Route path="/register" element={<CreateUser />}></Route>
         </Routes>
       <Footer />
     </div>
